scripts/requirements.txt
appdirs==1.4.4
astroid==2.4.2
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
ptyprocess==0.7.0
Pygments==2.7.4
pylint==2.6.0
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==21.0.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
